tidying dockerfile and requirements
