+++ Dockerfile
@@ -1,22 +1,6 @@
-FROM python:3.9-slim-buster
+FROM python:3.8
 
 
-# Install netcat
-RUN apt-get update && apt-get install -y netcat
-RUN pip install -U flask
-RUN pip install kafka-python
-# RUN pip install confluent-kafka
 
-WORKDIR /app
-
-COPY requirements.txt .
-
-RUN pip install -r requirements.txt
-
-COPY . .
-
-EXPOSE 8080
-
-CMD ["flask", "run", "--host=0.0.0.0", "--port=8080"]
-
-
+COPY requirements.txt requirements.txt
+RUN pip3.8 install -r requirements.txt
