+++ app.py
@@ -5,6 +5,7 @@ from kafka import KafkaProducer
 from kafka.errors import NoBrokersAvailable
 import time
 import json
+from kafka_helpers import check_connection_status, get_producer
 
 app = Flask(__name__)
 
@@ -12,51 +13,9 @@ app = Flask(__name__)
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 
-def check_connection_status(host, port):
-    s = socket.socket()
-    try:
-        logger.info(f"Checking {host} connection status")
-        s.connect((host, port))
-    except Exception as e:
-        s.close()
-        logger.info(f"No, {host} connection is not yet open. Retry in 5sec.")
-        time.sleep(5)
-        return check_connection_status(host, port)
-    logger.info(f"Ok, {host} connection is open!")
-    s.close()
-
-print('Hello')
-
-# time.sleep(20)
-
-# Check Kafka connection status
+#Kafka details
 host = 'kafka'
 port = 9093
-producer = None
-# check_connection_status(host, port)
-
-# Set up Kafka producer
-# producer = KafkaProducer(
-#     bootstrap_servers='kafka:9092',  # Replace with your Kafka broker address
-#     value_serializer=lambda v: json.dumps(v).encode('utf-8')
-# )
-
-def create_kafka_producer(host, port):
-    check_connection_status(host, port)
-    return KafkaProducer(
-        bootstrap_servers=f'{host}:{port}',
-        value_serializer=lambda v: json.dumps(v).encode('utf-8')
-    )
-
-def get_producer():
-    global producer
-    while producer is None:
-        try:
-            producer = create_kafka_producer(host, port)
-        except NoBrokersAvailable:
-            logger.info("Retrying to create Kafka producer in 5 seconds...")
-            time.sleep(5)
-    return producer
 
 @app.route('/', methods=['GET'])
 def sendid():
@@ -65,17 +24,6 @@ def sendid():
     logger.info(f"User-Id: {user_id}")
     logger.info(f"Session-Id: {session_id}")
     return '241'
-# def hello_world():
-#     prod = get_producer()
-#     if prod is not None:
-#         message = {"text": "Hello, World!"}
-#         prod.send('test', message)
-#         return 'Hello, World!'
-#     else:
-#         return 'Error: No Brokers Available', 500
-# def log_request():
-#     Log incoming request
-#     print(f"h")
 
 @app.route('/evt', methods=['POST'])
 def evt():
