+++ kafka_consumer.py
@@ -1,8 +1,9 @@
 from kafka import KafkaConsumer
+from kafka_helpers import check_connection_status, get_producer
 
 # Replace 'localhost:9092' with the appropriate broker address
 consumer = KafkaConsumer(
-    'test', # Replace with your topic name
+    'evt', # Replace with your topic name
     bootstrap_servers=['kafka:9093'],
     auto_offset_reset='earliest',
     value_deserializer=lambda x: x.decode('utf-8')
