+++ requirements.txt
@@ -1,1 +1,2 @@
 Flask
+kafka-python
